+++ demo/src/app/lists/table-list.component.ts
@@ -41,6 +41,7 @@ export class TableListComponent implements OnChanges {
 
   @Input()
   items: ListItem[];
+  scrollItems: ListItem[];
   indices: any;
 
   filteredList: ListItem[];