+++ demo/src/app/lists/vertical-list.component.ts
@@ -1,5 +1,5 @@
-import { Component, Input, OnChanges } from '@angular/core';
-
+import { Component, Input, OnChanges, ViewChild } from '@angular/core';
+import { VirtualScrollComponent } from 'angular2-virtual-scroll';
 import { ListItem } from './list-item.component';
 
 @Component({
@@ -10,6 +10,7 @@ import { ListItem } from './list-item.component';
     <button (click)="reduceListToEmpty()">Reduce to 0 Items</button>
     <button (click)="reduceList()">Reduce to 100 Items</button>
     <button (click)="setToFullList()">Revert to 1000 Items</button>
+    <button (click)="scrollTo()">Scroll to 50</button>
 
     <div class="status">
         Showing <span class="badge">{{indices?.start + 1}}</span>
@@ -32,9 +33,13 @@ export class VerticalListComponent implements OnChanges {
 
   @Input()
   items: ListItem[];
+  scrollItems: ListItem[];
   indices: any;
 
   filteredList: ListItem[];
+
+  @ViewChild(VirtualScrollComponent)
+  private virtualScroll: VirtualScrollComponent;
 
   reduceListToEmpty() {
     this.filteredList = [];
@@ -54,6 +59,10 @@ export class VerticalListComponent implements OnChanges {
 
   setToFullList() {
     this.filteredList = (this.items || []).slice();
+  }
+
+  scrollTo() {
+    this.virtualScroll.scrollInto(this.items[50]);
   }
 
   ngOnChanges() {